Separate out wdl kubernetes test to avoid missing dependency
